chore: update navigation and status bar styles
This commit updates the navigation and status bar styles in the App.js file. The status bar is set to use the light style, and the navigation stack is configured with custom header and content styles. The header background color is changed to "#36273C", the header text color is set to white, and the content background color is set to "#503E57". The "MealCategories" screen is given a title of "All Categories", and the "MealsOverview" screen has a custom header back title and style.